requirements.txt
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
coverage==7.2.7
requests==2.31.0
setuptools==65.5.0